+++ deepchecks/vision/metrics_utils/detection_precision_recall.py
@@ -19,6 +19,8 @@ import torch
 from ignite.metrics import Metric
 from ignite.metrics.metric import reinit__is_reduced, sync_all_reduce
 
+from deepchecks.vision.metrics_utils.metric_mixin import MetricMixin
+
 
 def _dict_conc(test_list):
     result = defaultdict(list)
@@ -35,7 +37,7 @@ def _dict_conc(test_list):
     return result
 
 
-class AveragePrecisionRecall(Metric):
+class AveragePrecisionRecall(Metric, MetricMixin):
     """Abstract class to calculate average precision and recall for various vision tasks.
 
     Parameters
@@ -334,28 +336,3 @@ class AveragePrecisionRecall(Metric):
             if zeroed_negative:
                 res = res.clip(min=0)
             return res[0][0]
-
-    @abstractmethod
-    def get_confidences(self, detections) -> List[float]:
-        """Get detections object of single image and should return confidence for each detection."""
-        pass
-
-    @abstractmethod
-    def calc_pairwise_ious(self, detections, labels) -> Dict[int, np.ndarray]:
-        """Get single result from group_class_detection_label and return matrix of IOUs."""
-        pass
-
-    @abstractmethod
-    def group_class_detection_label(self, detections, labels) -> dict:
-        """Group detection and labels in dict of format {class_id: {'detected' [...], 'ground_truth': [...]}}."""
-        pass
-
-    @abstractmethod
-    def get_detection_areas(self, detections) -> List[int]:
-        """Get detection object of single image and should return area for each detection."""
-        pass
-
-    @abstractmethod
-    def get_labels_areas(self, labels) -> List[int]:
-        """Get labels object of single image and should return area for each label."""
-        pass
